requirements.txt
asgiref==3.6.0
Django==4.2
psycopg2==2.9.6
PyJWT==2.6.0
python-dotenv==1.0.0
sqlparse==0.4.3
tzdata==2023.3